requirements.txt
django==3.0.7
djangochannelsrestframework==0.1.0
